users dashboard index

+++ app/Http/Controllers/BookController.php
@@ -10,6 +10,7 @@ use App\Models\Category;
 use App\Models\Publisher;
 use App\Models\BookCategory;
 use App\Models\User;
+use App\Models\BookLocation;
 use GuzzleHttp\Psr7\Query;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -70,7 +71,7 @@ class BookController extends Controller
     /**
      * Store book
      */
-    public function storeBook(BookStoreRequest $request, Author $author,  Publisher $publisher, Category $category){
+    public function storeBook(BookStoreRequest $request, Author $author,  Publisher $publisher, Category $category, BookLocation $bookLocation){
 
         $author = $this->checkAuthor($request, $author);
         $publisher = $this->checkPublisher($request, $publisher);
@@ -90,6 +91,13 @@ class BookController extends Controller
             'publication_year' => $request->publication_year,
             'copies' => $request->copies ?? '1',
 
+        ]);
+
+        $bookLocation = BookLocation::create([
+            'book_id' => $book->id,
+            'call_number' => $request->call_number,
+            'floor' => $request->floor,
+            'shelf' => $request->shelf,
         ]);
 
         if($request->filled('category')) {
@@ -125,13 +133,13 @@ class BookController extends Controller
     /**
      * Update book details
      */
-    public function updateBook(BookUpdateRequest $request, Book $book, Author $author, Publisher $publisher, Category $category){
-
+    public function updateBook(BookUpdateRequest $request, Book $book, Author $author, Publisher $publisher, Category $category, BookLocation $bookLocation){
+       
         $author = $this->checkAuthor($request, $author);
         $publisher = $this->checkPublisher($request, $publisher);
         $category = $this->checkCategory($request, $category);
        
-       $book->update([
+        $book->update([
             'title' => $request->title,
             'author_id' => $author->id,
             'isbn' => $request->isbn,
@@ -140,7 +148,13 @@ class BookController extends Controller
             'copies' => $request->copies,
             'status' => $request->status
         ]);
-        
+        $bookLocation = BookLocation::where('book_id', $book->id);
+        $bookLocation->update([
+            'call_number' => $request->call_number,
+            'floor' => $request->floor,
+            'shelf' => $request->shelf,
+        ]);
+
         if($request->filled('category')) {
             if(BookCategory::where('book_id', $book->id)
                 ->where('category_id', $category->id)->count() == 0){
@@ -155,9 +169,11 @@ class BookController extends Controller
         
         return Redirect::route('book.edit', $book)->with('status', 'Book Info Updated Successfully');
     }
+ 
+    
 
     /**
-     * Check Author if exists and create if not
+     * Update Book Cover Photo
      */
     public function updateCover(BookUpdateRequest $request, Book $book){
 
@@ -177,26 +193,46 @@ class BookController extends Controller
     /**
      * Delete a Book 
      */
-    public function destroyBook(Request $request, Book $book){
+    public function destroyBook(Request $request, Book $book, Author $author, Publisher $publisher){
         
         $validated = Validator::make($request->all(), [
             'password' => ['required', 'current_password']
-           ]);
-    
-           if($validated->fails()){
+        ]);
+
+        $author = Author::where('id', $book->author_id)->first();
+        $publisher = Publisher::where('id', $book->publisher_id)->first();
+      
+        if($validated->fails()){
            
             return Redirect::back()->with('book', "$book->id")->withErrors($validated);
             
-           }else{
+        }else{
 
             $imagePath = 'storage/book_cover/'. $book->cover;
             if(File::exists($imagePath)){
                 File::delete($imagePath);
             }
-                $book->delete();
+            // Book Destroy
+            $book->delete();
+            
+            //Check Author and Publisher before deleting
+            if($author->books->count() === 0 && $publisher->books->count() === 0){
+                $author->delete();
+                $publisher->delete();
+                return Redirect::route('book.index')->with('status', 'Book, Author and Publisher Deleted Successfully');
+            }elseif($author->books->count() === 0 && $publisher->books->count() != 0){
+                $author->delete();
+                return Redirect::route('book.index')->with('status', 'Book and Author Deleted Successfully');
+            }elseif($author->books->count() != 0 && $publisher->books->count() === 0){
+                $publisher->delete();
+                return Redirect::route('book.index')->with('status', 'Book and Publisher Deleted Successfully');
+            }else{
                 return Redirect::route('book.index')->with('status', 'Book Deleted Successfully');
-           }
+            }
+        }
 
+          
+       
     }
 
     /**
@@ -240,7 +276,9 @@ class BookController extends Controller
         return $category;
     }
 
-
+    /**
+     * Display all authors and searched results
+     */
     public function authors(Request $request) {
         $authors = Author::with('books')->withCount('books')->orderBy('name', 'asc')->paginate(10);
         if($search = $request->search){
@@ -257,10 +295,16 @@ class BookController extends Controller
         return view('book.admin.authors', compact('authors'));
     }   
 
+    /**
+     * Edit Author Name
+     */
     public function authorEdit(Author $author) {
         return view('book.admin.author-edit', compact('author'));
     }
 
+    /**
+     * Update Author Name
+     */
     public function authorUpdate(Request $request, Author $author) {
         $validated = $request->validate(['name' => ['string', 'unique:authors,name,'.$author->id]]);
         $author->update($validated);
@@ -268,12 +312,22 @@ class BookController extends Controller
        return Redirect::route('author.edit', $author)->with('status', 'Author Updated Successfully');
     }
 
+    /**
+     * Destroy Author
+     */
     public function authorDestroy(Author $author) {
-        $author->delete();
-
-        return Redirect::route('book.authors');
+        if($author->books->count() === 0){
+            $author->delete();
+            return Redirect::route('book.authors')->with('status', 'Author Deleted Successfully');
+        }else{
+            return Redirect::back()->with('statusError', 'Author Cannot Be Deleted Because It Has Book');
+        }
+       
     }
     
+    /**
+     * Display all publishers and searched results
+     */
     public function publishers(Request $request) {
 
         $publishers = Publisher::with('books')->withCount('books')->orderBy('name', 'asc')->paginate(10);
@@ -291,10 +345,16 @@ class BookController extends Controller
         return view('book.admin.publishers', compact('publishers'));
     }
 
+    /**
+     * Edit Publisher Name
+     */
     public function publisherEdit(Publisher $publisher) {
         return view('book.admin.publisher-edit', compact('publisher'));
     }
 
+    /**
+     * Update Publisher Name
+     */
     public function publisherUpdate(Request $request, Publisher $publisher) {
         $validated = $request->validate(['name' => ['string', 'unique:publishers,name,'.$publisher->id]]);
         $publisher->update($validated);
@@ -302,12 +362,23 @@ class BookController extends Controller
        return Redirect::route('publisher.edit', $publisher)->with('status', 'Publisher Updated Successfully');
     }
 
+    /**
+     * Destroy Publisher Name
+     */
     public function publisherDestroy(Publisher $publisher) {
-        $publisher->delete();
+        
+        if($publisher->books->count() === 0){
+            $publisher->delete();
+            return Redirect::route('book.publishers')->with('status', 'Publisher Deleted Successfully');
 
-        return Redirect::route('book.publishers');
+        }else{
+            return Redirect::back()->with('statusError', 'Publisher Cannot Be Deleted Because It Has Book');
+        }
     }
 
+    /**
+     * Display All Categories and Search Result 
+     */
     public function categories(Request $request) {
         $categories = Category::orderBy('name', 'asc')->paginate(10);
         if($search = $request->search){
@@ -324,11 +395,17 @@ class BookController extends Controller
         return view('book.admin.categories', compact('categories'));
     }
 
+    /**
+     * Display Create Category Page 
+     */
     public function createCategory() {
     
         return view('book.admin.category-create');
     }
 
+    /**
+     * Store Category
+     */
     public function storeCategory(Request $request, Category $category){
 
         $request->validate([
@@ -341,12 +418,18 @@ class BookController extends Controller
 
         return Redirect::route('category.create')->with('status', 'New Category Created');
     }
-
+    
+    /**
+     * Display Selected Category to Edit  
+     */
     public function categoryEdit(Category $category) {
 
         return view('book.admin.category-edit', compact('category'));
     }
 
+    /**
+    * Update Selected Category  
+    */
     public function categoryUpdate(Request $request, Category $category) {
         $validated = $request->validate(['name' => ['string']]);
         $category->update($validated);
@@ -354,12 +437,18 @@ class BookController extends Controller
        return Redirect::route('category.edit', $category)->with('status', 'Category Updated Successfully');
     }
 
+    /**
+    * Delete/Destroy Selected Category  
+    */
     public function categoryDestroy(Category $category) {
         $category->delete();
 
-        return Redirect::route('book.categories');
+        return Redirect::route('book.categories')->with('status', 'Category Deleted Successfully');
     }
 
+    /**
+    * Delete/Destroy Category added to a book  
+    */
     public function bookCategoryDestroy(Book $book,BookCategory $bookCategory) {
         $bookCategory->delete();
